pyproject.toml
[project]
name = "sentinel-sre"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "huggingface-hub>=1.8.0",
    "openenv-core>=0.2.2",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "ruff>=0.15.9",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"] # Ignore long lines for SRE logs and prompts
